:zap: Drop renderSize change for entering edit mode
I think this is overall faster, as if the user has established in their context that they wish to edit a specific field (with index say 1001) and double clicked on it, then a loading spinner while we render every field up to that point seems appropriate. A longer term fix is to allow the user to project out just that field (plus the _id field) and enable field-based updates.