COMPASS-1787 <ExpansionBar> in edit mode for >99% speedup of click Edit button (#19)
* :white_check_mark: Add disableHideButton flag test

* :tada: Initial disableHideButton flag implementation

Decided we don't want to handle the case of a modified field being hidden off-screen - what should happen?
Does the modification get applied or rejected?

* :zap: Remove edit/delete special case

These should now work correctly, in theory...

* :zap: For edit/update success/cancel/delete, reset to INITIAL_FIELD_LIMIT

Avoids rendering dozens/hundreds of fields over again. The user can always click to re-expand the document if they have reason to not be confident the update has gone through.

* :shirt:

* :zap::memo: Apply perClickSize performance prop, improve comments

* :fire: Can't edit in delete mode, so we can go faster

* :zap: Drop renderSize change for entering edit mode

I think this is overall faster, as if the user has established in their context that they wish to edit a specific field (with index say 1001) and double clicked on it, then a loading spinner while we render every field up to that point seems appropriate. A longer term fix is to allow the user to project out just that field (plus the _id field) and enable field-based updates.

* :memo: Add link explaining future work

* :bug: Scroll the top of the current document into view on cancel, update, delete

* :ok_hand::hammer: Drop overzealous guards

* :ok_hand::memo: Rename everythingSize -> totalSize in test names