Remove closeAllMenus action:
The action was only used in 1 place, the LineNumber component, and its
purpose was to ensure that only 1 contextual menu from this component
was open at any given time. Since we also had react-click-outside in
place, that module was handling this as well.

Since there was no need for this action, it is now removed. All
behaviour checks out as correct after the change.